Correct zfs send -v progress
It turns out only one ioctl may be active at anytime on one open file
descriptor, so the send-progress thread needs to reopen the /dev/zfs
node.

Also added async code to handle IOCTL that way, by returning
STATUS_PENDING, but in the send case did not fix the progress issue
anyway, and due to its complicated nature, it is commented-out.

+++ ZFSin/spl/include/sys/file.h
@@ -10,7 +10,7 @@
 struct spl_fileproc {
     void        *f_vnode;  // this points to the "fd" so we can look it up.
     list_node_t  f_next;   /* next zfsdev_state_t link */
-    int          f_fd;
+    uint64_t     f_fd;
     uint64_t     f_offset;
     void        *f_proc;
     void        *f_fp;
@@ -21,8 +21,8 @@ struct spl_fileproc {
 //typedef struct spl_fileproc file_t;
 #define file_t struct spl_fileproc
 
-void *getf(int fd);
-void releasef(int fd);
+void *getf(uint64_t fd);
+void releasef(uint64_t fd);
 /* O3X extended - get vnode from previos getf() */
 struct vnode *getf_vnode(void *fp);
